refactor connection_buffers into separate module
